Move unit tests to net462
