+++ UnitTests/UnitTests.csproj
@@ -1,9 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-
+    <TargetFramework>net462</TargetFramework>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
